+++ packages/widget-core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @agent-devtools/widget-core
 
+## 1.3.2-beta.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @agent-devtools/core@1.3.2-beta.0
+
 ## 1.3.1
 
 ### Patch Changes